+++ sdk/src/rhesis/sdk/metrics/config/backends.yml
@@ -3,7 +3,7 @@ backends:
     module: "rhesis.sdk.metrics.providers.deepeval"
     factory: "DeepEvalMetricFactory"
   ragas:
-    module: "rhesis.sdk.metrics.ragas"
+    module: "rhesis.sdk.metrics.providers.deepeval"
     factory: "RagasMetricFactory"
   rhesis:
     module: "rhesis.sdk.metrics.rhesis"
