+++ sdk/src/rhesis/sdk/metrics/providers/ragas/metrics.py
@@ -1,7 +1,7 @@
 from typing import List, Optional
 
 from rhesis.sdk.metrics.base import MetricResult, retry_evaluation
-from rhesis.sdk.metrics.ragas.metric_base import RagasMetricBase
+from rhesis.sdk.metrics.providers.ragas.metric_base import RagasMetricBase
 
 
 class RagasAnswerRelevancy(RagasMetricBase):
