+++ sdk/src/rhesis/sdk/metrics/tests/test_ragas.py
@@ -1,6 +1,6 @@
 import pytest
 
-from rhesis.sdk.metrics.ragas import RagasAnswerRelevancy, RagasContextualPrecision
+from rhesis.sdk.metrics.providers.ragas import RagasAnswerRelevancy, RagasContextualPrecision
 
 
 @pytest.fixture
